Minor correction and clarification
This guide suggested a 2.5A power supply for the raspberry pi 4 (probably as a carryover from when it used to reference a raspberry pi 3B+). However, the official documentation suggests at least a 3A PSU for the pi 4 (https://www.raspberrypi.org/documentation/hardware/raspberrypi/power/README.md)

Also added clarification that the raspberry pi itself does not show a user interface if plugging in a display after seeing multiple posts on discord/the forum with confusion surrounding this.